add try except to clients and fix some typing bugs

+++ weather_api/repos/weather/database.py
@@ -18,7 +18,7 @@ class WeatherAlchemyRepo(WeatherRepo):
             city: str,
             tmp: float,
             feels_like: float,
-            pressure: int,
+            pressure: float,
             humidity: int,
             wind_speed: float
             ) -> None:
